test: Remove 'ethereumj' dependency usages and fix more unsuccessful tests

+++ test-clients/src/main/java/com/hedera/services/bdd/suites/contract/opcodes/SStoreSuite.java
@@ -139,7 +139,7 @@ public class SStoreSuite extends HapiApiSuite {
 							for (int i = 0; i < numberOfIterations; i++) {
 								final var subOp1 = contractCall(contract, "changeArray",
 										//TODO : Integer -> BigInteger
-										ThreadLocalRandom.current().nextInt(1000))
+										BigInteger.valueOf(ThreadLocalRandom.current().nextInt(1000)))
 										.logged();
 								subOps.add(subOp1);
 							}
